// src/Types/PeopleManagement.tsx
export type TCertTitle =
    | '高空作業車操作人員'
    | '乙級職業安全管理員'
    | '甲級職業安全管理師'
    | '甲級職業衛生管理師'
    | '急救人員'
    | '缺氧(侷限)作業主管證照'
    | '有機溶劑作業主管證照'
    | '施工架組配作業主管證照'
    | '營造作業主管證照'
    | '作業主管證照'
    | '屋頂作業主管'
    | '鋼構組配作業主管'
    | '模板支撐作業主管'
    | '露天開挖作業主管'
    | '擋土支撐作業主管'
    | '粉塵作業主管';

export type TCertCode =
    | 'a'
    | 'bosh'
    | 'aos'
    | 'aoh'
    | 'fr'
    | 'o2'
    | 'os'
    | 'sa'
    | 's'
    | 'ma'
    | 'r'
    | 'ssa'
    | 'fs'
    | 'pe'
    | 'rs'
    | 'dw';